backend/api/requirements.txt
asgiref==3.2.10
blinker==1.4
Brotli==1.0.9
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cycler==0.10.0
dataclasses==0.6
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
Flask==1.1.2
h11==0.10.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kaitaistruct==0.8
kiwisolver==1.2.0
ldap3==2.7
MarkupSafe==1.1.1
matplotlib==3.3.1
mitmproxy==5.2
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.2
passlib==1.7.2
Pillow==7.2.0
protobuf==3.11.3
publicsuffix2==2.20191221
pyasn1==0.4.8
pycparser==2.20
pydivert==2.1.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.5.2
seaborn==0.10.1
six==1.15.0
sortedcontainers==2.1.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.10
urwid==2.0.1
Werkzeug==1.0.1
wsproto==0.15.0
zstandard==0.13.0
